Refactor: Move link-markdown component to prompt-kit directory
