+++ components/prompt-kit/markdown.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { LinkMarkdown } from "@/app/components/chat/link-markdown"
+import { LinkMarkdown } from "./link-markdown"
 import { cn } from "@/lib/utils"
 import { marked } from "marked"
 import { memo, useId, useMemo } from "react"
